+++ actionpack/CHANGELOG.md
@@ -1,10 +1,14 @@
+*   Show development mode exceptions with alert dialog for JavaScript requests.
+
+    *Patrik Bóna*
+
 *   Use accept header in integration tests with `as: :json`
 
     Instead of appending the `format` to the request path. Rails will figure
     out the format from the header instead.
 
     This allows devs to use `:as` on routes that don't have a format.
-    
+
     Fixes #27144.
 
     *Kasper Timm Hansen*